==> method_A/04.1_validate_fine-mapping.Rmd <==
---
title: "Untitled"
author: "Jason Torres"
date: "11/23/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)
library("tidyverse")
library("GenomicRanges")
library("viridis")

serv.dir <- "/Users/jtorres/FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"

work.dir <- proj.dir %&% "method_A/"
analysis.dir <- work.dir %&% "analysis_files/"
plot.dir <- work.dir %&% "plots/"

keep.df <- fread(proj.dir %&% "input_data/" %&% "380.locus.ID.txt")
keep.vec <- keep.df$Locus.ID_inCREDS

block.df <- fread(work.dir %&% "multi_results/results_blocks.txt")

```


Read data files 

```{r}

block.df <- fread(work.dir %&%"multi_results/results_blocks.txt",sep="\t")
fcred.df <- fread(work.dir %&%"multi_results/results_func-cred-sets.txt",sep="\t")
full.df <- fread(analysis.dir  %&% "classified_full.txt")

```



# Validation 2 - Comparison with islet only credible sets 

Build analysis data frame 

```{r}

get_max_over_null <- function(id,query.df,null.df){
  q <- max(filter(query.df,Locus.ID==id)$PPA)
  n <- max(filter(null.df,Locus.ID==id)$PPA)
  dif <- q - n 
}

get_num_vs_null <- function(id,query.df,null.df){
  q <- length(filter(query.df,Locus.ID==id)$PPA)
  n <- length(filter(null.df,Locus.ID==id)$PPA)
  dif <- q - n 
}

build_comp_df <- function(group.df){
  fcred.multi.df <- fread(work.dir%&%"multi_results/results_func-cred-sets.txt",sep="\t")
  fcred.islet.df <- fread(work.dir%&%"islet_results/results_func-cred-sets.txt",sep="\t")
  fcred.null.df <- fread(work.dir%&%"null_results/null_results_func-cred-sets.txt",sep="\t")
  Locus.ID <- unique(fcred.multi.df$Locus.ID)
  islet.max.vs.null <- map(Locus.ID,function(id){
    q <- max(filter(fcred.islet.df,Locus.ID==id)$PPA)
    n <- max(filter(fcred.null.df,Locus.ID==id)$PPA)
    dif <- q - n 
  }) %>% as.numeric(.)
  islet.num.vs.null <- map(Locus.ID,function(id){
    q <- length(filter(fcred.islet.df,Locus.ID==id)$PPA)
    n <- length(filter(fcred.null.df,Locus.ID==id)$PPA)
    dif <- q - n 
  }) %>% as.numeric(.)  
  multi.max.vs.null <- map(Locus.ID,function(id){
    q <- max(filter(fcred.multi.df,Locus.ID==id)$PPA)
    n <- max(filter(fcred.null.df,Locus.ID==id)$PPA)
    dif <- q - n 
  }) %>% as.numeric(.)
  multi.num.vs.null <- map(Locus.ID,function(id){
    q <- length(filter(fcred.multi.df,Locus.ID==id)$PPA)
    n <- length(filter(fcred.null.df,Locus.ID==id)$PPA)
    dif <- q - n 
  }) %>% as.numeric(.) 
  out.df <- data.frame(Locus.ID,islet.max.vs.null,islet.num.vs.null,
                       multi.max.vs.null,multi.num.vs.null,stringsAsFactors=FALSE)
  out.df <- filter(out.df,Locus.ID %in% keep.vec)
  out.df$refseq <- map(out.df$Locus.ID,function(id){filter(block.df,Locus.ID==id)$refseq}) %>% 
    as.character(.)
  out.df$group <- map(out.df$Locus.ID,function(id){filter(group.df,Locus.ID==id)$assigned}) %>% 
    as.character(.)
  return(out.df)
}


build_test_df <- function(comp.df,group.df){
  grps <- unique(group.df$assigned) 
  grps <- (grps)[!(grps %in% c("NA"))]
  Group <- c()
  w.pval.max <- c()
  w.pval.num <- c()

  for (g in grps){
    sub.df <- filter(comp.df,group==g)
    pval1 <- wilcox.test(sub.df$islet.max.vs.null,sub.df$multi.max.vs.null)$p.value 
    pval2 <- wilcox.test(sub.df$islet.num.vs.null,sub.df$multi.num.vs.null)$p.value 
    Group <- c(Group,g)
    w.pval.max <- c(w.pval.max,pval1)
    w.pval.num <- c(w.pval.num,pval2)    
  }
  sub.df <- filter(comp.df,group!="islet"&group!="unclassified")
  pval1 <- wilcox.test(sub.df$islet.max.vs.null,sub.df$multi.max.vs.null)$p.value 
  pval2 <- wilcox.test(sub.df$islet.num.vs.null,sub.df$multi.num.vs.null)$p.value 
  Group <- c(Group,"peripheral")
  w.pval.max <- c(w.pval.max,pval1)
  w.pval.num <- c(w.pval.num,pval2)  
  out.df <- data.frame(Group,w.pval.max,w.pval.num,stringsAsFactors=FALSE)
  return(out.df)
}

make_test_plots <- function(group.df){
  comp.df <- build_comp_df(group.df)
  test.df <- build_test_df(comp.df,group.df)
  test.df$Group <- factor(test.df$Group,levels=c("islet",
                                                 "liver",
                                                 "muscle",
                                                 "adipose",
                                                 "peripheral",
                                                 "unclassified"))
  #col.vec <- c("olivedrab2","gold1","grey","brown","red","ghostwhite")
  plt1 <- ggplot(data=test.df,aes(x=Group,y=-log(w.pval.max,base=10))) + 
    geom_point(shape=21,color="black",
               fill="grey",size=2) + 
    scale_y_continuous(limits=c(0,3),breaks=seq(0,3,0.5)) + 
    theme_bw() + geom_hline(yintercept=-log(0.05,base=10),
                            color="firebrick3",linetype=2) + 
    ylab(expression(-log[10](p[value]))) + 
    ggtitle("Wilcoxon test for difference in improvement of max PPA over null",
            subtitle="Mean of Islet-only fgwas = " %&% round(mean(grp1$islet.max.vs.null),5) %&% 
            "\nMean of multi-tissue fgwas = " %&% round(mean(grp1$multi.max.vs.null),5))
    
  plt2 <- ggplot(data=test.df,aes(x=Group,y=-log(w.pval.num,base=10))) + 
    geom_point(shape=21,color="black",
               fill=brewer.pal(11,"Spectral")[9],size=2) + 
    scale_y_continuous(limits=c(0,3),breaks=seq(0,3,0.5)) + 
    theme_bw() + geom_hline(yintercept=-log(0.05,base=10),
                            color="firebrick3",linetype=2) + 
    ylab(expression(-log[10](p[value]))) + 
    ggtitle("Wilcoxon test for difference in change of credible set size vs null",
            subtitle="Median of Islet-only fgwas = " %&% round(median(grp1$islet.num.vs.null),1) %&% 
            "\nMedian of multi-tissue fgwas = " %&% round(median(grp1$multi.num.vs.null),1))
  return(list(plt1,plt2))
}



```


# Statistical tests 

```{r}

grp1 <- filter(comp.df,group=="islet")
shapiro.test(grp1$islet.max.vs.null) # p-value = 9.507e-05
shapiro.test(grp1$multi.max.vs.null) # p-value = 1.413e-10
# Data is not likely normally distributed 
wilcox.test(grp1$islet.max.vs.null,grp1$multi.max.vs.null) 


grp2 <- filter(comp.df,group!="islet")
shapiro.test(grp2$islet.max.vs.null) # p-value = 3.693e-08
shapiro.test(grp2$multi.max.vs.null) # p-value = 3.112e-12
# Data is not likely normally distributed 



plt.list <- make_test_plots(full.df)
  

ggsave(plot=plt.list[[1]],filename=plot.dir%&%"Wilcox-compare-max_full.png",height=3,width=7)
ggsave(plot=plt.list[[2]],filename=plot.dir%&%"Wilcox-compare-num_full.png",height=3,width=7)

```
